fix(xdata): TransactionResult is UInt8, not signed int8
The decode was casting to int8_t, making 0x80 (tecCLAIM=128) appear
as -128, 0x96 (tecKILLED=150) as -106, etc. These are valid tec
codes that exist on-chain. Keep as unsigned uint8 → int32.

Also remove unnecessary _str_to_enum monkey-patch from verifier.

+++ src/x-data/includes/catl/xdata/codecs/enum.h
@@ -168,8 +168,7 @@ struct TransactionResultCodec
     decode(Slice const& data, Protocol const& protocol)
     {
         uint8_t raw = data.data()[0];
-        // TransactionResult codes are signed
-        int32_t code = static_cast<int8_t>(raw);
+        int32_t code = static_cast<int32_t>(raw);
         for (auto const& [name, c] : protocol.transactionResults())
         {
             if (c == code)
